backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]
